A bunch of fixes. See extended description.
Add regex for new API output, adding Cookie value for further optimizations, and implementing isBlocked to fix construct returning true issue.

For further optimizations, if user is not blocked they too will be given a cookie to remove duplicating ipinfo requests if user reloads page. 

You can fix this by using this script in a sort-of "Gateway mode" whereas the page determines if the user is blocked or not and redirects instead of displaying content on the same page. However, a fix will be implemented soon.